renamed:    api/index.ts -> src/api/index.ts     	renamed:    components/calendar/Calendar.tsx -> src/components/calendar/Calendar.tsx 	renamed:    components/calendar/common/CalendarItem.tsx -> src/components/calendar/common/CalendarItem.tsx 	renamed:    components/calendar/common/DateInfo.tsx -> src/components/calendar/common/DateInfo.tsx 	renamed:    components/carousel/carousel.tsx -> src/components/carousel/carousel.tsx 	renamed:    components/footer/Footer.tsx -> src/components/footer/Footer.tsx 	renamed:    components/statsList/List.tsx -> src/components/statsList/List.tsx 	renamed:    components/statsList/Stats.tsx -> src/components/statsList/Stats.tsx 	renamed:    config/defineLocale.ts -> src/config/defineLocale.ts 	renamed:    data/data.ts -> src/data/data.ts 	renamed:    pages/Calendar/index.tsx -> src/pages/Calendar/index.tsx 	renamed:    pages/Home/index.tsx -> src/pages/Home/index.tsx 	renamed:    redux/dateSlice.ts -> src/redux/dateSlice.ts 	renamed:    redux/statsSlice.ts -> src/redux/statsSlice.ts 	renamed:    redux/store.ts -> src/redux/store.ts
