+++ src/components/carousel/carousel.tsx
