+++ src/components/statsList/List.tsx
